pyproject.toml
[project]
name = "automate-test-appium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appium-python-client>=5.2.7",
    "pytest>=9.0.2",
    "selenium>=4.41.0",
]
